# Implement parallel file IO for streamerless download