--- a/test/test.cpp
+++ b/test/test.cpp
@@ -84,7 +84,7 @@ BOOST_AUTO_TEST_CASE(IOThreadPoolTest)
 	}
 	for (size_t i = 0; i < queue.size(); ++i) {
 		int val = queue[i];
-		pool.submit(handlers[val], [&counters, &total_ret_counter, val] () -> IOThreadPool::OptRetF {
+		handlers[val].submit([&counters, &total_ret_counter, val] () -> IOThreadPool::OptRetF {
 			counters[val] += 1;
 			if (val % modRetEval == 0) {
 				return [&total_ret_counter] () {