The main goal is to improve download performance on Windows as there is a lot of overhead when creating
many small files caused by antivirus software.

It improves download of `byar:test` on my internet connection:
- Linux: ~18.5s -> ~13.5s
- Windows: ~50s -> ~22s

~I still have some profiling and tuning to do to see what is the bottleneck on Windows and if we can make it as performant as on Linux, but I would like to submit this bulk of code already for the review.~

Tested some simple ideas, but not much gained. Overall even when I drop File IO altogether, Windows still takes ~18s, so looks like curl itself is less optimized for that platform.

---
#### Implementation

100% of file operations and hashing of content to verify that file content is correct is moved to a dedicated thread pool. Each file belongs to a single random thread. Tasks on the thread pool are only accessing the file object, hash object, and io failure flag. File operation errors are communicated asynchronously back to the main thread that will then cancel the download.