@lhog 2022-08-24
The main engine uses `moodycamel` to put together ThreadPool. Is there any chance to harmonize with the engine implementation instead? Otherwise summoning @marcushutchings for his expert view on MT.