@p2004a 2022-08-24
I see engine is using move MPMC queue, I'm using simple SPSC variant, with queue per thread (it's good enough here) The thread pool here is very, very simple. The one in engine is more complex with a different feature set, but looking at the code, I could probably make it work here.

The thing is I don't think it's easily feasible to make that thread pool available here. Probably possible with a bunch of ifdefs and cmake magic but I'm not sure if it's worth it?