@lhog 2022-08-25
It's ok, it was just a question. If bringing engine constructs is more troublesome than valuable then by all means don't hesitate to continue with your own implementation

In general I'm sure you tested this one quite extensively so go ahead and merge it. There's currently no one around knowing pr-d better than you do. Own it :)