@p2004a 2022-08-25
> It's ok, it was just a question

Of course, just tried to answer it well :smile:.

Overall I like to have somebody to at least skim what I'm doing to avoid some stupid mistakes ;)